move ContractUtil to java binding
