+++ language-support/java/bindings-rxjava/src/main/java/com/daml/ledger/rxjava/grpc/ActiveContractClientImpl.java
@@ -6,9 +6,12 @@ package com.daml.ledger.rxjava.grpc;
 import com.daml.grpc.adapter.ExecutionSequencerFactory;
 import com.daml.ledger.api.v1.ActiveContractsServiceGrpc;
 import com.daml.ledger.api.v1.ActiveContractsServiceOuterClass;
-import com.daml.ledger.javaapi.data.*;
+import com.daml.ledger.javaapi.data.ActiveContracts;
+import com.daml.ledger.javaapi.data.ContractUtil;
+import com.daml.ledger.javaapi.data.GetActiveContractsRequest;
+import com.daml.ledger.javaapi.data.GetActiveContractsResponse;
+import com.daml.ledger.javaapi.data.TransactionFilter;
 import com.daml.ledger.rxjava.ActiveContractsClient;
-import com.daml.ledger.rxjava.ContractUtil;
 import com.daml.ledger.rxjava.grpc.helpers.StubHelper;
 import com.daml.ledger.rxjava.util.ClientPublisherFlowable;
 import io.grpc.Channel;
