+++ language-support/java/bindings/src/main/java/com/daml/ledger/javaapi/data/ContractUtil.java
@@ -1,9 +1,8 @@
 // Copyright (c) 2022 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.daml.ledger.rxjava;
+package com.daml.ledger.javaapi.data;
 
-import com.daml.ledger.javaapi.data.*;
 import com.daml.ledger.javaapi.data.codegen.Contract;
 import com.daml.ledger.javaapi.data.codegen.ContractCompanion;
 import com.daml.ledger.javaapi.data.codegen.InterfaceCompanion;
